feat: storage operations
This commit adds a thin abstraction layer on top of etcd operations to
provide a simpler interface that handles marshalling and unmarshalling
with type safety. It also adds a suite of narrow integration tests that
exercise the operations with an embedded etcd. I'm going to submit a
store that uses these operations in a separate PR.