feat: workflows etcd backend
This commit adds a custom Etcd backend for the go-workflows library.
This implementation should be compliant and functional, but see the
note in etcd_test.go for an issue I ran into with the E2E tests
provided by the library.

The other TODO in this commit is that one of the go-workflows packages
that I needed is marked as internal. The current state is that there is
an edge case where we won't be able to properly bubble up an error if
a sub workflow gets created twice. As far as I can tell, this shouldn't
happen unless there's a bug in the library, so I consider it a
non-blocking issue.

See the go-workflows documentation for details on how this will be used:
https://github.com/cschleiden/go-workflows