feat: add watch operation (#5)
The new WatchOp enables callers to watch a key or prefix for
modifications. Callers specify a callback function that takes an Event,
which contains details of the modification. WatchOp supports two
flavors of watch operations: WatchOp.Watch is non-blocking and intended
for persistent watches and WatchOp.Until blocks until the handler
returns true or until the specified timeout has elapsed.