ci: submodule test and lint support
This commit adds a Makefile with different test and lint recipes for
local and CI use. It also updates the circleci config to use the new
`ci` recipe that tests and lints all modules and outputs the results
in junit xml format.

The linter found some issues in the storage tests, so I resolved them
in this commit.