refactor: server module and ci improvements (#6)
Moves the server code to a `server` submodule and initializes a go
workspace. We intend to create multiple modules (server, client,
supporting libraries, etc.) in this project.

Also adds a Makefile with different test and lint recipes for local
and CI use. It also updates the circleci config to use the new `ci`
recipe that tests and lints all modules and outputs the results
in junit xml format.

The linter found some issues in the storage tests, so I resolved them
in this commit.