feat: api generation with goa
This commit adds an `api` module that uses Goa to generate HTTP client
and server code. The API specification so far is incomplete but does
have some of the core entities and operations.

Some of the generated examples in the OpenAPI spec are wonky, such as
repeating the same example for the `ArrayOf` properties. Hopefully I
can fix these in a followup commit.