feat(server): run command
This commit adds a `run` command to the server module that starts an
HTTP server and an HTTP-over-MQTT server. It also includes a
configuration mechanism that supports configuration via a file,
environment variables, or command line arguments. The HTTP server is
enabled by default in this configuration. The MQTT server is disabled
by default, but can optionally be enabled via a config file or
environment variables.