fix: add arbitrary opts, remove repo id constraint
This commit adds fields for arbitrary repository options for both
backups and restores. Users can use these fields to specify options that
are otherwise unsupported by our API.

This commit also remove the UUID constraint on the repo ID fields in the
backups API. These IDs are only used to compute the repo path, so it's
unnecessary to enforce that they are UUIDs.

This commit also removes the backup config ID field, which is unused and
unnecessary now that the backup config fields are single-valued.

Part of #16