feat: resource model
This commit adds a "resource model" to the control plane server that
enables us to represent the lifecycle and dependencies of each database
component (e.g. a Docker service or a Spock subscription) in a
consistent and generic way. This has many benefits, including making it
easier to make each database operation idempotent.

[PLAT-7]