feat: api improvements for backups and replicas (#21)
* feat: api improvements for backups and replicas

This commit makes a few minor improvements to the API for backups and
replicas.

Previously, the database spec could take a list of backup configs, each
with a list of nodes to which they apply. Backup configs with an empty
node list were interpreted as the default configuration. Now, the
database spec takes a single backup config, and I've removed the list of
nodes. Instead, the backup configuration can be overridden on each node.
This makes the API and the request processing much simpler and more
consistent, at the cost of making it more verbose to define some
configurations, for example, applying one config to 2/3 nodes.

```javascript
// before
{
    "backup_configs": [
        {
            "id": "default"
            // ...
        },
        {
            "id": "n1-backups",
            "nodes": ["n1"]
            // ...
        }
    ]
    // ...
}
// after
{
    "backup_config": {
        "id": "default"
        // ...
    },
    "nodes": [
        {
            "backup_config": {
                "id": "n1-backups"
                // ...
            }
            // ...
        }
    ]
    // ...
}
```

Previously, replicas were explicitly specified on each node in the spec.
This would have made the state confusing if the primary failed over to
one of its replicas. Now, I've made `host_ids` into a multi-valued
field, where the first host in the list is used as the primary during
the creation process, but otherwise, there's no distinction between
primaries and replicas in the spec. This makes the spec easier to reason
about regardless of the current state of the node's Patroni cluster.

```javascript
// before
{
    "nodes": [
	    {
		    "name": "n1",
		    "host_id": "host-1",
		    "read_replicas": [
			    {
				    "host_id": "host-2"
				},
				{
					"host_id": "host-3"
				}
			]
	    }
    ]
}
// after
{
    "nodes": [
        {
            "name": "n1",
            "host_ids": ["host-1", "host-2", "host-3"]
        }
    ]
    // ...
}
```

Part of #16.

* fix: add arbitrary opts, remove repo id constraint

This commit adds fields for arbitrary repository options for both
backups and restores. Users can use these fields to specify options that
are otherwise unsupported by our API.

This commit also remove the UUID constraint on the repo ID fields in the
backups API. These IDs are only used to compute the repo path, so it's
unnecessary to enforce that they are UUIDs.

This commit also removes the backup config ID field, which is unused and
unnecessary now that the backup config fields are single-valued.

Part of #16