feat: initialize pgbackrest and enable archiving (#23)
Configures pgBackRest, initializes the backup repository, and sets the
`archive_command` in the Patroni config.

For each node with a backup configuration, we're currently picking the
last host in the list to run backups. I have an in-progress refactor
that improves this mechanism so that it takes the current instance
state into account.

I've changed the container user from `pgedge` to `postgres` because the
upstream packages assume the `postgres` user is used to run Postgres
and any associated utilities, which means that directories they create
are owned by `postgres`. I ran into this issue with pgBackRest, which
is why I've included this change in this commit.

I considered that we could mitigate this issue in our images by
`chown`ing every directory, but it would be a hinderence for customers
who want to customize our image. By using the `postgres` user, upstream
packages work immediately.

Closes #17.