feat: improved change detection
Previously, we were just doing a byte-level diff of the JSON
representation of the current and desired state of a resource to
determine if an update was needed. This was flawed because some of the
resource properties are only known after a create/update/refresh,
similar to Terraform's "computed" attributes.

This commit adds a JSON diffing library that enables us to ignore the
computed resource attributes so that we're only comparing the known
values.

I've also added a "resource version" concept that enables us to force
a resource update even if the attributes haven't changed. This will be
useful if we need to make changes to resource logic that only modifies
the output and not any of the attributes.

PLAT-7