feat: create from backup
This feature enables you to create one or more nodes from an existing
pgBackRest backup by using the `restore_config` field on either the
database spec (to apply it to all nodes) or the node spec (to apply it
to a single node). This feature uses Patroni's "boostrap" functionality,
so the restore only happens when the node is created. Changing the
restore_config for existing nodes will affect the pgBackRest and
Patroni configuration, but the restore process will not run again.

In-place backups will build on this work in a subsequent PR.

Example create database request body that uses this feature for all
nodes:

```json
{
  "id": "5acc3f42-e24d-41ff-b487-37720f456a2e",
  "spec": {
    "database_name": "my_app2",
    "database_users": [
      {
        "username": "admin",
        "password": "password",
        "attributes": ["SUPERUSER", "LOGIN"]
      },
      {
        "username": "app",
        "password": "password",
        "attributes": ["LOGIN"],
        "roles": ["pgedge_application"]
      }
    ],
    "restore_config": {
      "provider": "pgbackrest",
      "source_database_id": "a93e2624-989f-4cad-b5cb-ac589c204713",
      "source_node_name": "n1",
      "source_database_name": "my_app",
      "repository": {
        "id": "4ad0674f-da63-439a-8db1-5d8cd34a947f",
        "type": "s3",
        "s3_bucket": "pgedge-backups-9f81786f-373b-4ff2-afee-e054a06a96f1",
        "s3_region": "us-east-1",
        "s3_key": "XXXXXXXXXXXXX",
        "s3_key_secret": "XXXXXXXXXXXXX"
      },
      "restore_options": [
        "--set=20250505-133225F",
        "--type=time",
        "--target=2025-05-05 13:34:00+00"
      ]
    },
    "nodes": [
      { "name": "n1", "host_ids": ["8f6e5455-e228-4e2e-9129-a86cba1437c8"] },
      { "name": "n2", "host_ids": ["36dcd7ff-9f04-476e-ac6f-5495d075607d"] },
      { "name": "n3", "host_ids": ["fa461a39-5867-4a72-9923-dd7ce91e1eab"] }
    ]
  }
}
```

PLAT-35