fix: error when backup already in progress
This commit improves the error behavior when a user attempts to initiate
a backup when a backup is already in progress for the given node.
Previously, this resulted in a 500 error and the associated task was
left in a `pending` state. Now, we delete the task and return a 409
error with a user-friendly message.

PLAT-32