refactor: database/utils.go improvements
This improves the efficiency of `WaitForPatroniRunning` and
`GetPrimaryInstanceID` by replacing `time.Sleep` with a `time.Ticker`.
This changes the behavior slightly so that if the context is cancelled
or timed out, each function will return immediately, whereas before it
could block for a little longer if the cancel ocurred during the
`time.Sleep`.

It also makes a small readability improvement to `GetPrimaryInstanceID`
to collapse the two conditions into one.

PLAT-35