refactor: improve PgBackRestRestore efficiency
Previously, this function was processing some arguments more than once
when it didn't need to. I included some tests to validate that the core
functionality is unchanged. I changed `PgBackRestRestoreCmd` to public
because we prefer external test packages where possible. I also changed
the visibility of `PgBackRestBackupCmd` for symmetry.

PLAT-35