feat: ecr repository for control-plane image
This commit slighty modifies our ECR repository setup and adds a new
deployment for the `pgedge-assets` account.

The image repos module was not providing any value over just using the
ECR repository resource directly, so I've removed it to simplify the
setup. I've also added a repository for the control plane image.
Finally, I've added `make` recipes to deploy the Terraform backend and
configuration to make it easier to deploy.

PLAT-54