fix: remove backup config from restored nodes
Restoring an instance from backup can change the system ID. PgBackRest
stanzas are tied to a specific system ID, so the stanza needs to be
recreated on a new repository when the system ID changes. In order to
provide safe and consistent behavior, we will remove the backup
configuration from any nodes that are being restored. Users can then
submit an update request with a new repository after the restore has
completed. Users will see the updated spec with the backup configuration
removed in the initiate-database-backup response, as well as any other
endpoints that return the database spec.

PLAT-56