fix: instability in instance updates
This adds a short initial in `WaitForService` and adds some additional
checks to verify that the service has hit its desired state. See the
inline comment for more details about this change.

It also adds some tolerance for transient connection errors from the
Patroni API. This lengthens the time to detect a failed instance in
exchange for more durability when the Patroni reload operation causes
a brief unavailibility.

PLAT-98