fix: delay resuming workflows after restart
The `go-workflows` backend locks workflows and activities while they're
being executed. In the event of a crash or restart, the lock would
prevent the worker from resuming the workflow. This fix makes a few
changes to resolve this issue:

- Uses a stable ID for the worker ID
- Adds a random "worker instance" ID
- Adds the new IDs to the lock entities
- Incorporates the new IDs into the lock checks

If the worker ID matches the lock's worker ID but the worker instance ID
does not match, we know that the lock belonged to an old instance of the
worker and that the lock can be reassigned to the new instance.

PLAT-99