ci: improve go caching (#45)
Adds the go build cache directory to our saved cache so that we won't
rebuild the go tool binaries unless there are module updates in the
subsequent steps.

It also adds fallbacks to the restore_cache step so that it will restore
from the most recent cache when we've modified the go dependencies or
tools.

This change reduces the run time of our CI jobs by a few minutes.