feat: api versioning
This commit adds versioning to our API endpoints, with the current
version as `v1`. Under this scheme, each API version lives under
separate packages and is served under separate top-level prefixes, e.g.
`/v1` for the current version.

This necessitated a small refactor to make it easier to modularize the
initialization logic.

PLAT-86