chore: add api docs server to control-plane-dev
Adds an API docs server to the `control-plane-dev` configuration. You'll
be able to access the API documentation at `http://localhost:8999` when
you run `make dev-watch`.

Being able to view the API docs in this format makes it easier to review
some parts of the generated OpenAPI spec, such as request and response
examples.

PLAT-86