feat: allow human-readable IDs
This modifies our API contract to allow human-readable IDs, as opposed
to UUIDs, in each user-specified ID field:

- Cluster ID
- Host ID
- Database ID
- Tenant ID

Instance IDs are still deterministically generated from the host ID,
database ID, and node name, but are designed to be more readable. For
example, given these inputs:

- Host ID: "us-east-1"
- Database ID: "my-app"
- Node name: "n1"

We will generate the instance ID "my-app-n1-n5fe2mcy". We're using a
short hash of the host ID rather than the host ID itself to enhance
readability when the host IDs are long, such as those used by Cloud.
For example, when both the host ID and database ID are UUIDs:

- Host ID: "dbf5779c-492a-11f0-b11a-1b8cb15693a8"
- Database ID: "ed2362ea-492a-11f0-956c-9f2171e8b9ab"
- Node name: "n1"

The instance ID is "ed2362ea-492a-11f0-956c-9f2171e8b9ab-n1-io5979nh"

This commit contains breaking changes for the development environment,
so it's recommended to do the following before updating to this commit:

- Delete all databases via the API
- Remove existing data directories:
  `rm -rf docker/control-plane-dev/data`

This will reset your environment, so you'll need to reinitialize the
cluster.

PLAT-86