feat: censor sensitive fields in API responses
With this change, the following sensitive fields will be omitted from
all API responses:

- Database user passwords
- Backup and restore repository credentials:
  - S3 keys/secrets
  - Azure keys
  - GCS keys

We want to enable users to manage databases without storing the original
create request. So, this commit also changes the update behavior so that
these sensitive fields can be omitted from update requests. This makes
it safe to store the database spec in version control.

Users can modify the credentials, such as when rotating passwords, by
including the updated values in an update request.

PLAT-86