feat: surface volume validation errors
Builds on the existing volume validation to surface more detailed error
information, including the host, node, and underlying error message.

Also makes some minor changes that make it possible to distinguish
between invalid volumes and other types of errors. If an unexpected
error occurs, such as an image pull error, the user will get a 500
instead of a 400 error.

PLAT-86