feat: port conflict validation
Adapts the `ValidateInstanceSpec` activity to take and validate multiple
specs in order to detect when a port has been allocated multiple times
on a single host.

Note that this will not consistently detect when a port has been
allocated multiple times on a machine when there are multiple
control-plane instances with different host IDs on the same machine - as
in our development configuration. End users should not run the control
plane this way, so we donn't need to handle that case explicitly.

PLAT-86