fix: unnecessary stanza create events (#60)
The `pgbackrest info` output contains a non-zero status code if the
repository is empty, even if it's otherwise configured correctly. We
were returning a `resource.ErrNotFound` in this case, which caused us
to re-run the `PgBackRestStanza.Create` method unnecessarily. The
`Create` implementation is idempotent, so this didn't cause issues, but
it was confusing to see the create event in the output when there were
no changes.