feat: OpenAPI examples
Adds request and response examples for each endpoint in the OpenAPI
spec.

I had to modify our `list-*` endpoints to return an object rather than
an array in order to get the examples to generate properly. This is best
practice anyway, so it's a good overall change.

I also removed the `MaxLength` annotations from `g.MapOf` attributes
because this is not properly supported by Goa and it caused some
confusing output in the docs.

I recommend reviewing this PR with the local API docs server at
http://localhost:8999.

PLAT-120