fix: change etcd init log level
Some of the etcd settings we use require a restart, so when we
initialize the embedded etcd, we run a temporary instance first. Etcd
notifies us that it's ready to be used through its `ReadyNotify`
channel, but starting in etcd v3.6, not all of the initialization is
complete when this notification fires. As a result, we get some error
messages when we shut down the temporary etcd instance.

There's no way for us to reliably wait until the etcd has started all
of its handlers, so this change drops the log level to `fatal` to remove
those error logs. The permanent etcd instance still uses the
configurable log level.

PLAT-183