fix: block after init and join calls until ready (#82)
The control plane starts in an uninitialized state until you either
initialize a new cluster or join the instance to an existing cluster.
We've implemented this process using polymorphism. There are two
handler implementations: "pre-initialization" and "post-initialization".
When you initialize or join a cluster, we swap those implementations.
But, prior to this fix, we haven't always completed that swap when the
init or join call returns. If you're initializing a cluster from a
script, there's a high likelihood that the first call after init-cluster
will produce an error.

This adds a mechanism that lets us block until the handlers are swapped
before we return from either the init-cluster or join-cluster calls.
This makes it safe for clients to immediately join the next host to a
cluster after init-cluster has returned.

PLAT-187