fix: ignore module tags
Modifes the `Makefile` to ignore the module-specific tags we create
during the release. `goreleaser` has limited capabilities to filter
tags in the free version. Instead, we're able to set the current tag
explicitly via the `GORELEASER_CURRENT_TAG` environment variable.

We don't need to do this in the release job in CircleCI because it
checks out a specific tag and it doesn't fetch all other tags. So,
it only has the tag that we want.