feat: custom elector implementation
Adds a custom elector implementation that avoids the etcd concurrency
package, which has several outstanding issues on GitHub. It also avoids
using KeepAlive, because of a suspected goroutine leak.

Instead, it creates a new lease each time it refreshes the leader lock.
The leader will refresh the lock once every 10 seconds, so this is a
relatively small expense.

PLAT-182