fix: get primary instance activity host id
We were not using the hostID parameter to ExecuteGetPrimaryInstance.
This fixes a bug where we were unable to execute this activity depending
on the cluster topology and which server was running the calling
workflow.

I looked into adding the `unparam` linter to find similar issues, but
it wasn't possible to do without catching many false positives because
we need to inspect exported functions. Instead, we'll need to be
dilligent in code review to catch similar issues.

PLAT-225