fix: use unique workflow IDs (#94)
This changes our workflows to use unique, random instance IDs rather
than deterministic ones. We were using deterministic IDs as a safety
mechanism to prevent overlapping workflows in case the database status
check failed. But, the deterministic IDs cause problems for clients
because they don't have a way to validate that a workflow has completed
apart from the task status. We set the task status within the workflow,
which means that clients might see that the task is marked completed
before the workflow has exited. If they try to start another workflow
too quickly, they would get an error, even though the task appears
completed.

This change fixes that issue by allowing for overlapping workflows.
So, as soon as the database status is updated and a task is marked
completed, the client is immediately able to start another action.

PLAT-200