feat: cancel database task endpoint
This pr adds an implementation for an endpoint that allows users to cancel a task and its associated workflow.
Long running activities poll the task status and exit early when the task is cancelled.

In order to use this endpoint to cancel a task, specify
- a database id (example)
- an associated task id (0198ce28-23ad-71ec-8617-33dd1391cd4e)

restish control-plane-local-1 cancel-task example 0198ce28-23ad-71ec-8617-33dd1391cd4e

PLAT-160