fix: timing issue in ec2 fixtures
Fixes a timing issue when starting the ec2 fixtures while they're in a
stopped state. The `public_ip` property was consistently missing in this
scenario. This change uses the `ec2_instance_info` task to fetch the
instance info after the create, and this has worked reliably in my
testing.

PLAT-200