fix: temporary workaround for add-node
In order for `sub_create` to copy structure/data, the target database
needs to be completely empty. Snowflake and LOLOR both create schemas
when they're added to a database, so they can interfere with the
add-node process if we create them before we create subscriptions.

This commit is just a temporary fix to get add-node to run. We need to
move these `CREATE EXTENSION` statements to run after we run
`spock.sub_create`.

PLAT-143