fix: reuse endpoint between mqtt clients
MQTT brokers require client IDs to be unique per client. The broker will
send disconnect packets to connections with duplicate client IDs.

This change makes it possible to reuse an `mqtt.Endpoint` between
HTTP-over-MQTT clients, which makes it easier to configure a client for
multiple control plane servers.

NOTE: callers are responsible for calling Connect on the MQTT endpoint.
See the updates to the tests for the updated usage.