fix: error in pg_replication_slot_advance query
We removed the CTE from this query in a recent refactor, but forgot to
remove this reference. In most cases, these replication slots are
already at or ahead of where they need to be, so this query doesn't run
very often. That's why this bug evaded detection for so long.