chore: zsh dev-env utilities (#129)
Adds a `source`-able script with helper functions for interacting with
the control plane development environment with zsh. Also adds a plugin
and theme for Oh My Zsh that works with the helper functions.

See the included `hack/dev-env.md` file for instructions to use the
helper functions and the Oh My Zsh integration.