fix: idempotent pg_replication_slot_advance
We preemptively create replication slots for the disabled subscriptions
in the "populate node" process (also called Z0DAN). Before we enable
those subscriptions, we check the current position of the replication
slots against what's already been replicated to the new node. If the
slot is behind what's already been replicated, we call
`pg_replication_slot_advance` to skip the duplicate transactions.

But, this process involves queries that run across multiple instances.
When the replication slot is advancing normally, it can get ahead of
the LSN we give to `pg_replication_slot_advance` in the time it takes
us to run the requisite queries. When this happens, we get an error:

```
cannot advance replication slot to X, minimum is Y (SQLSTATE 55000)
```

This commit makes the `pg_replication_slot_advance` call conditional
within the query, which makes this operation idempotent and avoids
the error above.