fix: serve post-init api after initialization (#141)
Changes the startup order of the app services so that we don't serve the
post-init handlers until after every other initialization step. This
change also makes it so the join and init cluster endpoints will return
an error if an error occurs during the initialization process.

Since the server is now ready to take requests before the init or join
call returns, we no longer need the workaround in the multi-server
client that we added in #106.

PLAT-241