refactor: remove cohort id
We were using this property to differentiate groups of hosts that are
running in different Swarms or K8s clusters. However, this field is
unavailable on Swarm workers, and there are several other issues we'll
need to resolve to support heterogeneous clusters. We've decided to
revisit these issues when start adding support for other orchestrators.
Until then, we can simplify our code by removing this unused field.

PLAT-210