feat: random etcd password for hosts
Generates a random Etcd password for hosts and stores it in the
generated config file in the data directory. We generate the password
when we're creating the Etcd user for the new host, and we write it
to the config file in `writeCredentials` where we were already writing
the peer and client certificates for the new host. I've also added the
username and password to the internal `GetJoinOptions` endpoint
response. This is necessary for the `join-cluster` flow.

PLAT-105