feat: add support for generated config file (#145)
Adds a mechanism to read/write a generated config file in the data
directory. The new `config.Manager` manages the process of loading
the user-specified configuration and combining it with the defaults
and the generated config file with this preference in order of most-
preferred to least-preferred:

1. User-specified config
  - The order of preference amongst these configs is set by the order
    of `Sources`, which is defined in `server/cmd/root.go`.
2. Generated config
3. Defaults

This enables users to override settings in the generated config if
necessary.

PLAT-105