fix: set needs recreate on resource not found
Changes the behavior of the apply events activity so that rather than
removing the resource from the state, we set `NeedsRecreate = true` when
the resource's Refresh method returns a `resource.ErrNotFound`. This
fixes an issue where the first time setup logic can be triggered by
accident in some situations.

PLAT-308