fix: handle postgres service stopped
We were checking whether the Postgres service was stopped in the service
spec resource, and then restarting it if needed. This was problematic
because the spec resource executes on the instance host, which may not
be a Swarm manager.

This change removes that responsibility from the spec resource and
modifies the service resource to detect changes between the current and
desired spec - including whether the service needs to be scaled up.

Now, when the postgres service is scaled down, the service resource's
Refresh method will detect that an update is needed. The normal service
deploy method we use in the Create/Update methods will scale the service
back up.

PLAT-308