test: add pull-through cache to e2e fixtures
Adds a pull-through cache for ghcr.io to the E2E fixtures. This makes
the E2Es much more efficient because we only need to pull each image
from the upstream repository once, and then its pulled from the cache.
It solves the problem where every host is pulling the same images
simultaneously, which could easily hit bandwidth limits in the 12-host
configurations.

You can't push to a repository that's configured as a pull-through
cache, so we needed to use a separate repository from the one we use for
our control plane images.

PLAT-308