feat: host recovery (#204)
<!-- PR title format: follow Conventional Commits style, e.g.
     feat: add new postgres minor release images
     fix: correctly mark stale instances as unknown
     docs: add database migration guide
Common types: feat, fix, docs, chore, test, refactor, build(release)
     See: https://www.conventionalcommits.org/en/v1.0.0/#summary
-->

## Summary

This PR enhances the "remove-host" endpoint by adding a "force" option
which allows users to recover from situations where a host is lost.
Prior to this, calling "remove-host" on a lost host would fail because
the internal state would indicate a db running and prevent the removal.

## Changes

- Updated the API endpoint to:
  - include an optional "force" param
- return a `RemoveHostResponse` containing a slice of tasks related to
the host removal
  - propagated related changes through the client package  
- Updated logic in `post_init_handler.go` to bypass the instance check
during host removal
- Added new RemoveHost workflow/task that propagates the db changes
across the rest of the cluster
- Added logic to prevent additional requests being sent to the dead
host's work queue
- Added a new test to `clustertest` framework (more below)

## Testing

Added a new test case and utils to the `clustertest` framework:
- create a new 3 node cluster and database
- wait for it to become healthy
- forcibly kill one of the host containers
- call `remove-host --force` to remove the missing container from the
cluster state
- wait for it to become healthy
- assert that the cluster has the expected number of instances, nodes,
etc in the expected states

## Checklist

- [x] Tests added or updated (unit and/or e2e, as needed)
- [x] Documentation updated (if needed)
- [x] Issue is linked (branch name or URL in PR description)
- [x]
[Changelog](https://github.com/pgEdge/control-plane/blob/main/docs/development/development.md#pull-requests)
entry added for user-facing behavior changes
- [ ] Breaking changes (if any) are clearly called out in the PR
description

## Notes for Reviewers

There are a couple related follow-on tasks:
1. the current task management layer is db-specific and needs some
refactoring to allow tasks for additional entity types (like hosts):
https://pgedge.atlassian.net/browse/PLAT-347
2. there is an issue preventing cleanup of the dead instance from the
cluster after forcibly removing it. The cluster is healthy, but when you
call `get-database`, you can see that there is still an instance
leftover in the "unknown" state.
- as a result, there is a workaround func in utils that allows the test
to pass even though the extra zombie instance exists - this will need to
be cleaned up when the root issue is resolved.