fix: persist recovery opts in in-place restore
It's not possible to persist recovery options outside of the bootstrap
process and it's not possible to force Patroni into the bootstrap
process if the pgdata directory is non-empty.

So, this commit changes the in-place restore process to temporarily move
the pgdata directory to a different name. Then, it changes the bootstrap
command to simply move the pgdata directory back to its original name.
This lets us keep our existing restore process and makes patroni keep
the recovery options.

PLAT-368