test: verify restore to specific backup in backup / restore tests (#232)
## Summary

This PR improves the backup / restore tests by properly checking that we
can restore to a specific backup, rather than the only backup. When you
solely pass in `set` to pgBackRest with a backup label, the default
behavior of pgBackRest is to restore to that backup, and then recover to
the latest point in the archive. Setting `type` to `immediate` tells
Postgres to only recover until the database becomes consistent. The
previous test did not differentiate between these two scenarios.

## Checklist

- [X] Tests added or updated (unit and/or e2e, as needed)
- [X] Documentation updated (if needed)
- [X] Issue is linked (branch name or URL in PR description)
- [X]
[Changelog](https://github.com/pgEdge/control-plane/blob/main/docs/development/development.md#pull-requests)
entry added for user-facing behavior changes
- [X] Breaking changes (if any) are clearly called out in the PR
description