build(release): bump version to v0.6.2 (#234)
## Summary

Release PR for v0.6.2.

_Note: The legacy `tenant_id` field for Cluster was never populated (it
could not be set by clients) and never returned data. As such, we have
kept this a patch release despite this change in spec. Validations on
this field were preventing `GetCluster` calls in certain scenarios._

### Removed

- Removed unused "tenant_id" field from GetCluster endpoint.

### Fixed

- Fixed in-place restores where the recovery target is not the latest
point in the repository.

## Checklist

- [X] Tests added or updated (unit and/or e2e, as needed)
- [X] Documentation updated (if needed)
- [X] Issue is linked (branch name or URL in PR description)
- [X]
[Changelog](https://github.com/pgEdge/control-plane/blob/main/docs/development/development.md#pull-requests)
entry added for user-facing behavior changes
- [X] Breaking changes (if any) are clearly called out in the PR
description